Move trainer, evaluator, and generator to root
