+++ src/fairseq2/evaluator.py
@@ -7,16 +7,15 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from collections.abc import Sequence
+from collections.abc import MutableMapping, Sequence
 from contextlib import nullcontext
-from typing import Any, Final, TypeVar, final
+from typing import Any, Generic, TypeVar, final
 
 import torch
-from torch import Tensor
+from torch.nn import Module
 from torch.profiler import record_function
 from typing_extensions import override
 
-from fairseq2.checkpoint import CheckpointManager
 from fairseq2.data_type import DataType
 from fairseq2.datasets import DataReader
 from fairseq2.device import CPU, SupportsDeviceTransfer
@@ -25,56 +24,65 @@ from fairseq2.gang import GangError, Gangs, raise_operational_gang_error
 from fairseq2.logging import log
 from fairseq2.metrics import MetricBag, sync_and_compute_metrics
 from fairseq2.metrics.common import extend_batch_metric_values
-from fairseq2.metrics.recorders import (
-    NOOP_METRIC_DESCRIPTOR,
-    MetricDescriptor,
-    MetricRecorder,
-)
+from fairseq2.metrics.recorders import MetricRecorder
 from fairseq2.profilers import Profiler
-from fairseq2.recipe.evaluator import EvalUnit
+from fairseq2.recipe.model import RecipeModel
+from fairseq2.task import Task, TaskStopException
 from fairseq2.typing import ContextManager
 from fairseq2.utils.device_stat import DeviceStatTracker
 from fairseq2.utils.progress import ProgressReporter
 from fairseq2.utils.rng import RngBag
 from fairseq2.utils.stopwatch import Stopwatch
+from fairseq2.utils.warn import _warn_deprecated
+
+BatchT_contra = TypeVar(
+    "BatchT_contra", bound=SupportsDeviceTransfer, contravariant=True
+)
+
+
+class EvalUnit(ABC, Generic[BatchT_contra]):
+    @abstractmethod
+    def prepare_metric_bag(self, metric_bag: MetricBag) -> None: ...
+
+    def set_train_step_nr(self, step_nr: int) -> None:
+        pass
+
+    @abstractmethod
+    def process_batch(self, batch: BatchT_contra, metric_bag: MetricBag) -> None: ...
+
+    def finalize(self, metric_bag: MetricBag) -> None:
+        pass
+
+    def process_metric_values(self, values: MutableMapping[str, object]) -> None:
+        pass
+
+    @property
+    def name(self) -> str | None:
+        return None
+
+    @property
+    def model(self) -> RecipeModel:
+        _warn_deprecated("`EvalUnit.model` is deprecated and will be removed in v0.14.")
+
+        raise NotImplementedError()
+
 
 BatchT = TypeVar("BatchT", bound=SupportsDeviceTransfer)
 
 
-class Validator(ABC):
-    @abstractmethod
-    def run(self, train_step_nr: int) -> float | None: ...
-
-    @abstractmethod
-    def reset(self) -> None: ...
-
-
 @final
-class _NoopValidator(Validator):
-    @override
-    def run(self, train_step_nr: int) -> float | None:
-        pass
+class Evaluator(Task):
+    """Evaluates a machine learning model."""
 
-    @override
-    def reset(self) -> None:
-        pass
-
-
-NOOP_VALIDATOR: Final = _NoopValidator()
-
-
-@final
-class StandardValidator(Validator):
     def __init__(
         self,
         *,
+        model: Module,
         units: Sequence[EvalUnit[BatchT]],
         data_readers: Sequence[DataReader[BatchT]],
         gangs: Gangs,
         amp: bool,
         amp_dtype: DataType,
-        score_metric_descriptor: MetricDescriptor,
-        checkpoint_manager: CheckpointManager,
         metric_recorder: MetricRecorder,
         profiler: Profiler,
         device_stat_tracker: DeviceStatTracker,
@@ -82,136 +90,76 @@ class StandardValidator(Validator):
         progress_reporter: ProgressReporter,
         seed: int,
     ) -> None:
-        """
-        :param units: The evaluation units.
-        :param data_readers: The data readers of ``units``.
-        :param wall_watch: The stopwatch to track process wall-time.
-        :param amp: If ``True``, enables ``torch.amp``.
-        :param score_metric_descriptor: The descriptor of the metric to use for
-            score calculation.
-        :param seed: The random number generator seed.
-        """
-        self._step_nr = 0
-
         if len(units) == 0:
-            raise ValueError("`units` must contain at least one validation unit.")
+            raise ValueError("`units` must contain at least one evaluation unit.")
 
         if len(units) != len(data_readers):
             raise ValueError(
                 f"Number of data readers in `data_readers` must match the number of units in `units` ({len(units)}), but is {len(data_readers)} instead."
             )
 
+        rng_bag = RngBag.from_device_defaults(CPU, gangs.root.device)
+
+        self._model = model
         self._units = units
-
         self._data_readers = data_readers
-
         self._gangs = gangs
-
         self._amp = amp
-
         self._amp_dtype = amp_dtype
-
-        self._score_metric_descriptor = score_metric_descriptor
-
-        self._checkpoint_manager = checkpoint_manager
-
-        self._rng_bag = RngBag.from_device_defaults(CPU, gangs.root.device)
-
+        self._rng_bag = rng_bag
         self._metric_recorder = metric_recorder
-
         self._profiler = profiler
-
         self._device_stat_tracker = device_stat_tracker
-
         self._data_watch = Stopwatch()
-
-        self._compute_watch = Stopwatch(device=self._gangs.root.device)
-
+        self._compute_watch = Stopwatch(device=gangs.root.device)
         self._lapse_watch = Stopwatch()
-
         self._wall_watch = wall_watch
-
         self._progress_reporter = progress_reporter
-
         self._seed = seed
-
+        self._step_nr = 0
+        self._stop_requested = False
         self._num_batches_read = 0
-
         self._has_run = False
-
-        self._best_score = -torch.inf
-
-        self._best_step_nr = -1
 
     @override
     @torch.inference_mode()
-    def run(self, train_step_nr: int) -> float | None:
+    def run(self) -> None:
         if self._has_run:
-            raise InvalidOperationError("Validator has already been run.")
+            raise InvalidOperationError("Evaluator has already been run.")
 
         self._has_run = True
 
-        self._maybe_restore_best_score_and_step()
+        with self._progress_reporter:
+            with self._rng_bag.temporary_manual_seed(self._seed):
+                with self._profiler:
+                    self._do_run()
 
-        with self._rng_bag.temporary_manual_seed(self._seed):
-            with self._profiler:
-                return self._do_run(train_step_nr)
+        self._gangs.close()
 
-    def _maybe_restore_best_score_and_step(self) -> None:
-        if self._score_metric_descriptor is NOOP_METRIC_DESCRIPTOR:
-            return
-
-        if self._best_step_nr >= 0:
-            return
-
-        if self._gangs.root.rank == 0:
-            scores_and_steps = self._checkpoint_manager.load_scores()
-        else:
-            scores_and_steps = []
-
-        self._gangs.root.barrier()
-
-        if scores_and_steps:
-            self._best_score, self._best_step_nr = scores_and_steps[0]
-        else:
-            self._best_step_nr = 0
-
-    def _do_run(self, train_step_nr: int) -> float | None:
-        scores = []
+    def _do_run(self) -> None:
+        self._model.eval()
 
         for unit, data_reader in zip(self._units, self._data_readers):
             if unit.name:
-                log.info("Validating {}.", unit.name)
+                log.info("Evaluating {}.", unit.name)
 
-            score = self._run_unit(train_step_nr, unit, data_reader)
+            self._run_unit(unit, data_reader)
 
-            if score is not None:
-                scores.append(score)
-
-        score = self._compute_aggregated_score(scores)
-
-        self._update_best_score(train_step_nr, score)
-
-        return score
-
-    def _run_unit(
-        self, train_step_nr: int, unit: EvalUnit[Any], data_reader: DataReader[Any]
-    ) -> float | None:
+    def _run_unit(self, unit: EvalUnit[Any], data_reader: DataReader[Any]) -> None:
         metric_bag = MetricBag(device=self._gangs.root.device)
 
         unit.prepare_metric_bag(metric_bag)
 
-        progress_task = self._progress_reporter.create_task("valid", total=None)
+        progress_task = self._progress_reporter.create_task("eval", total=None)
 
         self._device_stat_tracker.reset()
 
         eod = False
 
         with progress_task, self._lapse_watch:
-            unit.set_train_step_nr(train_step_nr)
-
             while not eod:
-                self._checkpoint_manager.maybe_complete_save_operation()
+                if self._stop_requested:
+                    raise TaskStopException()
 
                 batches = self._read_next_batches(unit, data_reader)
                 if batches is None:
@@ -221,7 +169,7 @@ class StandardValidator(Validator):
 
                 self._step_nr += 1
 
-                progress_task.step(1)
+                progress_task.step()
 
                 with record_function(f"step_{self._step_nr}"):
                     self._run_step(unit, batches, metric_bag)
@@ -232,9 +180,7 @@ class StandardValidator(Validator):
                 with record_function("finalize"):
                     self._call_unit_finalize(unit, metric_bag)
 
-        metric_values = self._publish_metrics(train_step_nr, unit, metric_bag)
-
-        return self._maybe_get_unit_score(metric_values)
+        self._publish_metrics(unit, metric_bag)
 
     def _read_next_batches(
         self, unit: EvalUnit[Any], data_reader: DataReader[Any]
@@ -253,7 +199,7 @@ class StandardValidator(Validator):
     def _run_step(
         self, unit: EvalUnit[Any], batches: list[Any], metric_bag: MetricBag
     ) -> None:
-        log.debug("Running validation step {}.", self._step_nr)
+        log.debug("Running step {}.", self._step_nr)
 
         with self._compute_watch:
             batches.reverse()
@@ -288,10 +234,8 @@ class StandardValidator(Validator):
 
         return torch.autocast(device_type=device_type, dtype=self._amp_dtype)
 
-    def _publish_metrics(
-        self, train_step_nr: int, unit: EvalUnit[Any], metric_bag: MetricBag
-    ) -> dict[str, object]:
-        log.debug("Syncing validation metrics.")
+    def _publish_metrics(self, unit: EvalUnit[Any], metric_bag: MetricBag) -> None:
+        log.debug("Syncing evaluation metrics.")
 
         gangs = self._gangs
 
@@ -331,21 +275,16 @@ class StandardValidator(Validator):
 
             values["wall_time"] = self._wall_watch.get_elapsed_time()
 
-            category = "valid"
+            category = "eval"
 
-            if unit.name is not None:
+            if unit.name:
                 category = f"{category}/{unit.name}"
 
-            self._metric_recorder.record_metric_values(category, values, train_step_nr)
+            self._metric_recorder.record_metric_values(category, values)
 
         gangs.root.barrier()
 
         self._reset_lapse_state()
-
-        if values is None:
-            values = {}
-
-        return values
 
     def _reset_lapse_state(self) -> None:
         self._data_watch.reset()
@@ -358,84 +297,15 @@ class StandardValidator(Validator):
 
         self._num_batches_read = 0
 
-    def _maybe_get_unit_score(self, metric_values: dict[str, object]) -> float | None:
-        if self._score_metric_descriptor is NOOP_METRIC_DESCRIPTOR:
-            return None
-
-        if self._gangs.root.rank != 0:
-            return None
-
-        metric_value = metric_values.get(self._score_metric_descriptor.name)
-        if metric_value is None:
-            return None
-
-        if not isinstance(metric_value, (int, float, Tensor)):
-            raise InternalError(
-                f"Score metric value must be of type `{int}`, `{float}`, or `{Tensor}`, but is of type `{type(metric_value)}` instead."
-            )
-
-        score = float(metric_value)
-
-        if not self._score_metric_descriptor.higher_better:
-            score = -score
-
-        return score
-
-    def _compute_aggregated_score(self, scores: list[float]) -> float | None:
-        if self._score_metric_descriptor is NOOP_METRIC_DESCRIPTOR:
-            return None
-
-        if self._gangs.root.rank != 0:
-            return 0.0
-
-        if not scores:
-            raise InternalError(
-                "None of the evaluation units returned a score metric value."
-            )
-
-        return sum(scores)
-
-    def _update_best_score(self, train_step_nr: int, score: float | None) -> None:
-        if self._score_metric_descriptor is NOOP_METRIC_DESCRIPTOR:
-            return
-
-        if score is None:
-            raise InternalError("`score` is `None`.")
-
-        if self._gangs.root.rank != 0:
-            return
-
-        if score > self._best_score:
-            self._best_score, self._best_step_nr = score, train_step_nr
-
-        if not log.is_enabled_for_info():
-            return
-
-        best_score = self._best_score
-
-        metric_descriptor = self._score_metric_descriptor
-
-        if not metric_descriptor.higher_better:
-            score, best_score = -score, -best_score
-
-        v1 = metric_descriptor.formatter(score)
-        v2 = metric_descriptor.formatter(best_score)
-
-        log.info("Score - Metric: {} | Last: {} (step {}) | Best: {} (step {})", metric_descriptor.display_name, v1, train_step_nr, v2, self._best_step_nr)  # fmt: skip
+    @override
+    def request_stop(self) -> None:
+        self._stop_requested = True
 
     @override
-    def reset(self) -> None:
-        self._step_nr = 0
+    def close(self) -> None:
+        self._metric_recorder.close()
 
-        for data_reader in self._data_readers:
-            data_reader.reset()
-
-        self._data_watch.reset()
-
-        self._compute_watch.reset()
-
-        self._lapse_watch.reset()
-
-        self._num_batches_read = 0
-
-        self._has_run = False
+    @property
+    @override
+    def step_nr(self) -> int:
+        return self._step_nr
