+++ src/fairseq2/recipe/__init__.py
@@ -18,15 +18,19 @@ from fairseq2.recipe.cli import main as main
 from fairseq2.recipe.cli import register_cli_error as register_cli_error
 from fairseq2.recipe.cli import train_main as train_main
 from fairseq2.recipe.dataset import RecipeDataset as RecipeDataset
-from fairseq2.recipe.evaluator import Evaluator as Evaluator
-from fairseq2.recipe.evaluator import EvalUnit as EvalUnit
-from fairseq2.recipe.generator import Generator as Generator
-from fairseq2.recipe.generator import GeneratorUnit as GeneratorUnit
 from fairseq2.recipe.model import RecipeModel as RecipeModel
 from fairseq2.recipe.run import evaluate as evaluate
 from fairseq2.recipe.run import generate as generate
 from fairseq2.recipe.run import run as run
 from fairseq2.recipe.run import train as train
 from fairseq2.recipe.tokenizer import RecipeTokenizer as RecipeTokenizer
-from fairseq2.recipe.trainer import Trainer as Trainer
-from fairseq2.recipe.trainer import TrainUnit as TrainUnit
+
+# isort: split
+
+# TODO: Deprecated, will be removed in v0.14.
+from fairseq2.evaluator import Evaluator as Evaluator
+from fairseq2.evaluator import EvalUnit as EvalUnit
+from fairseq2.generator import Generator as Generator
+from fairseq2.generator import GeneratorUnit as GeneratorUnit
+from fairseq2.trainer import Trainer as Trainer
+from fairseq2.trainer import TrainUnit as TrainUnit
