+++ src/fairseq2/recipe/composition/generator.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from fairseq2.recipe.generator import Generator
+from fairseq2.generator import Generator
 from fairseq2.recipe.internal.generator import _GeneratorFactory
 from fairseq2.runtime.dependency import (
     DependencyContainer,
